requirements.txt
anyio==4.12.1
argon2-cffi-bindings==25.1.0
argon2-cffi==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
comm==0.2.3
contourpy==1.3.3
criterion==0.0.1a1
cuda-bindings==13.0.3
cuda-pathfinder==1.2.2
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.62.0
fqdn==1.5.1
fsspec==2025.12.0
h11==0.16.0
h5py>=3.16.0,<3.17
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython-pygments-lexers==1.1.1
ipython==9.11.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib>=1.5.3,<1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema-specifications==2025.9.1
jsonschema==4.26.0
jupyter-client==8.8.0
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server-terminals==0.5.4
jupyter-server==2.17.0
jupyter==1.1.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
jupyterlab==4.5.5
kiwisolver==1.5.0
lark==1.3.1
markupsafe==3.0.3
matplotlib-inline==0.2.1
matplotlib>=3.10.8,<3.11
mistune==3.2.0
mpmath==1.3.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook-shim==0.2.4
notebook==7.5.4
numpy>=2.4.3,<2.5
pandas>=3.0.1,<3.1
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pillow==12.0.0
pip==26.0.1
platformdirs==4.9.4
prometheus-client==0.24.1
prompt-toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==3.0
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pyyaml>=6.0.3,<6.1
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-learn>=1.8.0,<1.9
scipy>=1.17.1,<1.18
send2trash==2.1.0
setuptools==80.10.2
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl>=3.6.0,<3.7
tinycss2==1.4.0
torch>=2.10.0,<2.11
torchvision>=0.25.0,<0.26
tornado==6.5.5
tqdm>=4.67.3,<4.68
traitlets==5.14.3
triton==3.6.0
typing-extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.46.3
widgetsnbextension==4.0.15
